TYP: fix ``timedelta64`` and ``datetime64`` method declarations
